Allow more accuracy (0.5 interval) when marking skill cooldown, add a slight delay to avoid using skills while cooldown expires

+++ src/Settings.cs
@@ -190,7 +190,7 @@ public class HotkeySlot
 	/// The cooldown time, in seconds.
 	/// </summary>
 	[JsonProperty]
-	public int Cooldown { get; set; }
+	public float Cooldown { get; set; }
 }
 
 public enum KeybindAction
